Rename ReconnectStrategy to RetryStrategy. Add send as a super trait.
